change all size_t type to uint64_t

+++ include/block/MetaBlock.h
@@ -16,11 +16,11 @@ namespace Doopey {
       //Meta struture: name length + name + ctime + number of block + BlockID list
       bool setFileName(string& filename);
       bool setCreateTime(time_t ctime);
-      bool setFileSize(size_t size);
+      bool setFileSize(uint64_t size);
       bool addDataID(BlockID id);
 
       string getFileName() const;
-      size_t getFileSize() const;
+      uint64_t getFileSize() const;
       time_t getCreateTime() const;
       uint64_t getDataBlockNumber() const;
       BlockID getDataBlockID(const uint64_t number) const;
@@ -28,7 +28,7 @@ namespace Doopey {
       uint64_t* _nameLength;
       char* _filename;
       time_t* _ctime;
-      size_t* _size;
+      uint64_t* _size;
       uint64_t* _blockNumber;
       BlockID* _firstBlockID;
   }; // class Block
